be/requirements.txt
django-apscheduler==0.6.2
djangorestframework==3.16.0
python-binance==1.0.28
numpy==1.26.4
pandas==2.2.3
pandas_ta==0.3.14b0
psycopg2-binary==2.9.10
redis==5.2.1
requests==2.31.0
python-decouple==3.8